# events/api/swagger.yaml
openapi: 3.0.0
info:
  title: Events API
  version: 1.0.0
  description: API for events
  contact:
    name: Events API
    url: http://localhost:8080
paths:
  /events:
    get:
      summary: List events
      description: List events
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
    post:
      summary: Create event
      description: Create event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Event'
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '500':
          description: Internal server error
      

  /events/{id}:
    get:
      summary: Get event
      description: Get event
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          description: Event not found
    put:
      summary: Update event
      description: Update event
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          description: Event not found
    delete:
      summary: Delete event
      description: Delete event
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          description: Event not found

components:
  schemas:
    Event:
      type: object
      properties:
        id:
          type: string
        faculty_id:
          type: string
        title:
          type: string
        description:
          type: string
        start_date:
          type: string
        end_date:
          type: string
        is_finished:
          type: boolean
        rating:
          type: number